+++ cmake/treedata/common/pvr.txt
@@ -2,6 +2,7 @@ xbmc/pvr            pvr
 xbmc/pvr/addons     pvr/addons
 xbmc/pvr/channels   pvr/channels
 xbmc/pvr/dialogs    pvr/dialogs
+xbmc/pvr/epg        pvr/epg
 xbmc/pvr/recordings pvr/recordings
 xbmc/pvr/timers     pvr/timers
 xbmc/pvr/windows    pvr/windows
