+++ xbmc/GUIInfoManager.cpp
@@ -91,7 +91,7 @@
 #include "pvr/PVRManager.h"
 #include "pvr/channels/PVRChannelGroupsContainer.h"
 #include "pvr/channels/PVRRadioRDSInfoTag.h"
-#include "epg/EpgContainer.h"
+#include "pvr/epg/EpgContainer.h"
 #include "pvr/recordings/PVRRecording.h"
 
 #include "addons/AddonManager.h"
