+++ xbmc/addons/interfaces/PVR/AddonCallbacksPVR.cpp
@@ -24,13 +24,13 @@
 #include "Application.h"
 #include "cores/VideoPlayer/DVDDemuxers/DVDDemuxUtils.h"
 #include "dialogs/GUIDialogKaiToast.h"
-#include "epg/EpgContainer.h"
 #include "events/EventLog.h"
 #include "events/NotificationEvent.h"
 #include "pvr/PVRManager.h"
 #include "addons/PVRClient.h"
 #include "pvr/channels/PVRChannelGroupsContainer.h"
 #include "pvr/channels/PVRChannelGroupInternal.h"
+#include "pvr/epg/EpgContainer.h"
 #include "pvr/recordings/PVRRecordings.h"
 #include "pvr/timers/PVRTimers.h"
 #include "pvr/timers/PVRTimerInfoTag.h"
