+++ xbmc/pvr/dialogs/GUIDialogPVRGuideInfo.cpp
@@ -18,10 +18,13 @@
  *
  */
 
+#include "GUIDialogPVRGuideInfo.h"
+
+#include <utility>
+
 #include "Application.h"
 #include "dialogs/GUIDialogOK.h"
 #include "dialogs/GUIDialogYesNo.h"
-#include "epg/EpgInfoTag.h"
 #include "guilib/GUIWindowManager.h"
 #include "guilib/LocalizeStrings.h"
 #include "messaging/ApplicationMessenger.h"
@@ -33,12 +36,9 @@
 #include "pvr/PVRManager.h"
 #include "pvr/addons/PVRClients.h"
 #include "pvr/channels/PVRChannelGroupsContainer.h"
+#include "pvr/epg/EpgInfoTag.h"
 #include "pvr/timers/PVRTimerInfoTag.h"
 #include "pvr/windows/GUIWindowPVRSearch.h"
-
-#include "GUIDialogPVRGuideInfo.h"
-
-#include <utility>
 
 using namespace PVR;
 using namespace KODI::MESSAGING;
