+++ xbmc/pvr/epg/CMakeLists.txt
@@ -1,0 +1,13 @@
+set(SOURCES EpgContainer.cpp
+            Epg.cpp
+            EpgDatabase.cpp
+            EpgInfoTag.cpp
+            EpgSearchFilter.cpp)
+
+set(HEADERS Epg.h
+            EpgContainer.h
+            EpgDatabase.h
+            EpgInfoTag.h
+            EpgSearchFilter.h)
+
+core_add_library(pvr_epg)
