+++ xbmc/pvr/epg/EpgSearchFilter.cpp
