+++ xbmc/pvr/epg/EpgSearchFilter.h
