+++ xbmc/pvr/timers/PVRTimerInfoTag.cpp
@@ -18,10 +18,10 @@
  *
  */
 
+#include "PVRTimers.h"
+
 #include "ServiceBroker.h"
 #include "dialogs/GUIDialogOK.h"
-#include "epg/Epg.h"
-#include "epg/EpgContainer.h"
 #include "messaging/ApplicationMessenger.h"
 #include "messaging/helpers/DialogHelper.h"
 #include "settings/AdvancedSettings.h"
@@ -34,8 +34,8 @@
 #include "pvr/PVRManager.h"
 #include "pvr/addons/PVRClients.h"
 #include "pvr/channels/PVRChannelGroupsContainer.h"
-
-#include "PVRTimers.h"
+#include "pvr/epg/Epg.h"
+#include "pvr/epg/EpgContainer.h"
 
 using namespace PVR;
 using namespace KODI::MESSAGING;
